fix: keep Firefox saves on the polyfill's blob download
Pulling correctly was necessary but not sufficient. With the writer
released, the polyfill's sink gets as far as posting the chunk and throws
"cannot transfer WebAssembly ArrayBuffer": StreamHelper.write hands it a
Uint8Array pointing straight into the wasm heap, and the sink puts
chunk.buffer in the transfer list. The file still arrives empty.

Both faults are inside Avalonia's bundled polyfill, so the streamed path
cannot be made to work from here. Its blob fallback has neither problem -
new Blob([chunk]) copies the bytes on the spot - and is what Safari has
always used. The polyfill takes it when getRegistration finds nothing, so
hide the registration from it.

Scoped to browsers without the native picker, leaving Chromium untouched.
Nothing else calls getRegistration: the offline cache and the update
prompt use the registration pwa.js already holds.

sw-file-save.js and the workers' half of the protocol go with it. They
only served a route that can never deliver bytes, and keeping them would
invite the polyfill back onto it.
