+++ src/CtrDxEditor.Browser/wwwroot/index.html
@@ -228,6 +228,7 @@
             </div>
         </dialog>
         <script type="module" src="./pwa.js"></script>
+        <script type="module" src="./save-fallback.js"></script>
         <script>
             // Hide the splash once Avalonia mounts content into #out (typically a canvas). A
             // MutationObserver is used rather than awaiting main.js's module promise, because
