+++ src/CtrDxEditor.Browser/wwwroot/service-worker.js
@@ -1,16 +1,5 @@
-// Development build: no offline support, so a stale cache can never shadow a rebuild.
-// The published build swaps in service-worker.published.js.
+// Development build: no offline support, so a stale cache can never shadow a rebuild. The
+// published build swaps in service-worker.published.js.
 //
-// Avalonia's save-file polyfill streams through whatever service worker is registered, so even a
-// worker that caches nothing has to speak that protocol or "Save level as" breaks. See
-// sw-file-save.js.
-self.importScripts("./sw-file-save.js");
-
-self.addEventListener("message", (event) => offerFileSave(event.data));
-
-self.addEventListener("fetch", (event) => {
-    const response = respondToFileSave(event.request);
-    if (response !== undefined) {
-        event.respondWith(response);
-    }
-});
+// Deliberately empty. pwa.js registers a worker in every build so the registration and update
+// paths are exercised in development too; this one just has nothing to cache.
